- Reformatted the code (NLP_model.py merged with bitcoin_abuse.py) - Changed the name of the dataset class (ReportDataset to BADataset)
